fixed arc version not being logged correctly, fixed faulty assert, fixed dps.report upload error on parsing tresponse

+++ GW2EIDPSReport/DPSReportJsons/DPSReportUploadEncounterObject.cs
@@ -4,7 +4,7 @@ public class DPSReportUploadEncounterObject
 {
     public string UniqueId;
     public bool Success;
-    public long Duration;
+    public double Duration;
     public long CompDps;
     public int NumberOfPlayers;
     public int NumberOfGroups;
